distribution summary for doubles
This is a second attempt at a distribution summary
type that supports recording double values instead
of just longs. There are quite a few things that
depend on long values for the main branch including:

* servo
    * counters for total and total of squares expect
      amounts to be int
    * max gauge expects int
* metrics{2,3}
    * Histogram expects a long

This attempt creates a summary object that can be
registered as a gauge. The values will get reset on
read if enough time has passed. That means that unlike
StepLong based implementations for servo the values
are not precise for a given interval and if the
sampling frequency doesn't match some events could
get lost or counted multiple times.

Since it isn't controlled by the registry,
implementations cannot control the set of statistics
supplied, there will always be a fixed set of 4:
count, total, totalOfSquares, max.