split AutoBindSingleton to separate library
There are a number of cases where users want full
control over the injection without having to worry
about the classpath scanning magic of AutoBindSingleton.
This change splits the core plugin into a separate
library, spectator-nflx-plugin, that is a normal
singleton and can be explicitly injected.

The spectator-nflx library now has an AutoPlugin
class that just injects the singleton plugin to
provide automatic loading.