add versions plugin and update dependencies
For equalsverifier, staying with 1.4.1 because
upgrading this is causing an
IncompatibleClassChangeError from the tests for
AutoPlugin. I think it is due to asm issues as
equalsverifier uses asm5 directly but depends
on cglib which needs asm4. Not important enough
for me to spend more time investigating right now.

For rxjava staying on rc.8 until rc.9 is working
from the internal repo.