improve support for guice
Adds modules that can be used to easily inject
the registry with governator for running the
actual app or with a test module that creates
an isolated registry for unit tests.

If the GcLogger is started multiple times it
will now log instead of throw. That class needs
to be refactored, but that will be a separate
PR. For this update the focus is just on making
the module work.