fix sporadic gaps on gauges reported to servo
Gauges were getting passed to servo by creating an
instance of NumberGauge when getMonitors is called.
NumberGauge keeps a weak reference to the number so
it will not prevent collection. Since spectator doesn't
hold onto a reference to the number passed to the
gauge it can get collected before the value is actually
read.

This change adds a simple ServoGauge monitor type that
keeps a strong reference to the value to report.