add helper for adding appender to root logger
The spectator appender needs to get a registry and
the registry loading tries to log. This creates a
problem where the slf4j loggers created before log4j2
successfully configures don't work properly and we
lose the logging.

We can delay the registry hook, but this can lead to
recursive calls to the appender or a misleading logger
name if we configure the logging to not send the
spectator messages to the spectator appender.

To work around this there is now a helper that can be
called as part of the application init to add the
SpectatorAppender to the root logger. It will also
listen for configuration change events and re-add the
appender as needed.

```java
SpectatorAppender.addToRootLogger(
  Spectator.registry(),
  "spectator",  // Appender name
  false)        // Ignore exceptions?
```

If more flexibility is needed you can still specify
the appender in the log4j config, but you will be missing
certain log events from spectator.

Also note, that the registry should not return an
implementation of a counter that will log on increment
as this would create a loop.