fix bucket labels
For some values, e.g., `latency(3, TimeUnit.SECONDS)`, the
unit was wrong so we would get labels 0s, 1s, and 3s. The
values should have been 0375ms, 0750ms, 1500ms, and 3000ms.
This change fixes those ranges and adds a test case to
ensure we are getting the proper number of buckets across
the default formatters.