work-around for null context class loader
In some cases we are seeing:

```
Thread.currentThread().getContextClassLoader() == null
```

For now this workaround will use the classloader of
the local class. In some quick tests this works and
unblocks the teams seeing this problem. We probably
need to find a better solution longer term after we
understand why the context class loader is null.