add helpers for creating id from maps
The helpers make it easier to create an id using a
`Map<String, String>` as the initial or additional
tag list.

This change does break API compatibility with an
addition to the `Id` interface:

```
:spectator-api:checkCompatibility
ERROR: com/netflix/spectator/api/Id.java: The method withTags(java.util.Map) has been added
Error count: 1
Warning count: 0
:spectator-api:checkCompatibility FAILED
```

It isn't expected that this will create many issues,
but we'll use this as a bit of a test before other
desirable API cleanup efforts. The main fallout, if
any, is expected to be mismatched versions of
`spectator-api` and `spectator-reg-servo` jars in
the classpath.