switch to wrap the main registry
Before the TDigestRegistry would get included via
the service loader as part of the composite. While
convenient for testing that causes everything to
use the digest timers which is potentially expensive
on the client and server.

With this change the expectation is that for use
cases that need it the TDigestRegistry will be
injected directly. The old behavior can be simulated
by mapping the main Registry binding to TDigestRegistry.