use registry for creating the ids
Updates the Json utility class to use an injected registry
to create the ids rather than using `DefaultId` directly.

This change does break backwards compatibility of some of
the classes to pass in the registry. However, it is still
experimental and our use of those classes is only via tests
or via a guice module so it will not break common usage.