global registry is now a composite
The static global registry returned by `Spectator.globalRegistry()`
is now a composite. During application initialization a user can
add another registry to the composite to subscribe to activity
hitting the global registry. If nothing is added the global will
behave similar to the NoopRegistry.

The primary benefit of this change is to make it easier to work
with DI systems. The DI library can be used to setup the main
registry and if there is interest in the global data add it to
the global registry.

If the legacy behavior is needed the user can call:

```java
Spectator.initializeUsingServiceLoader();
```

That method will add the registries found using the service loader
to the global composite.

The `Spectator.registry()` call is deprecated because it still
returns the `ExtendedRegistry` class that is already deprecated
and will go away in a later release.