fix name mapping for metrics3 registry
The `.` was not getting added between tags when
creating the names leading to:

```
server.requestCountagent-safaristatus-400method-GET
```

rather than the expected:

```
server.requestCount.agent-safari.status-400.method-GET
```

Also fixes a NPE if normalize is called on an Id with
an empty tag list. A test case was added for gauges
that are not currently getting mapped properly, but it
is ignored until that problem is fixed in a later commit.