add counters to track number of samples dropped
The digest timer will drop samples if another thread already
has the lock. For artificial benchmarks of the timer
we see a high loss rate as threads that do not perform an update
will take a shorter amount of time. In practice, the timer update
is usually small compared to the code being measured.