refactor TagList to be more intuitive
Previous behavior was based on some tests to reduce the
cost of generating the dynamic ids in high volume paths
compared to servo. However, it led to a number of surprising
and confusing issues for users. With this change we want
to clear up the behavior for users and we can optimize
as needed in the future.

It will now sort by the tag key and remove duplicates on
update. The goal is to have more intuitive behavior for
the user so they can rely on:

1. Any ids with the same set of tags will be equal.
2. Later additions will always override if the key already
   exists. So for `withTags(Iterable)` if there were
   entries with duplicate keys the iteration order would
   determine which one ended up getting used.