add percentile timer and distribution summary
This is an alternative to using digests to get an approximate
percentile value that works well with slicing and dicing. For
this approach we are using a fixed set of buckets that partition
the range of long values so it can be used with most sets of
non-negative long values.

The error for an estimate is usually within a few percent of
the actual percentile for the real data sets tried, but there
are some exceptions. Compared with using t-Digests it is far
more performant and usually more compact. That comes at a
cost of being less accurate and more restrictive on inputs.

Docs will be updated after the cheaper bucket options are
moved from sandbox to stable. We do need to be careful with
the overhead and encourage users to be diligent about the
usage of percentiles and the overhead compared to options
such as a normal timer or bucket timer. A rule of thumb
is that a percentile timer will be ~ 300x more expensive in
terms of metrics volume than a normal timer.