delay reporting until there is activity
For the servo registry delay marking the counter, timer, or
distribution summary as active until there is some activity
such as a call to increment on the counter. This is mainly
for use-cases that cache the meter instance to avoid the
lookup cost, but may not get used for quite a while.