move bucket helpers from sandbox to api
These have been used for quite a while and are considered
stable now. The old classes in sandbox have been kept and
marked as deprecated to avoid breaking existing users. To
avoid drift in the implementations they now just wrap the
new versions in the api lib.

A few small changes as part of the move:

* Remove BucketFunction in favor of jdk LongFunction
* Accessor to get type for a given bucket changed to package
  private