indicate type as prefix on idx value
Timers are mapped to nanoseconds before computing the
bucket. This allows better precision without needing
to support floating point. For display it is often
preferred to use base units, like seconds. In any
case, it is useful to distinguish whether the value
was reported for a timer or a distribution summary. The
index value stored for the `percentile` key is now
prefixed with a `T` for timers and `D` for distribution
summaries.