try directory structure to make browsing easier
Github wiki had flat layout with page titles being mapped
to the file names. Since that is no longer a requirement
try creating a directory structure that is easier to
follow.

Note, need to be a bit careful here because the structure
impacts the links the user will have to use.