fixes #292, remove overload for ValueFunction
This will be a backwards incompatible change, but based
on depotsearch results should have minimal impact at
Netflix. All uses either explicitly cast or do not use
the variant that would cause problems.

This also updates the compatibility test jar with some
additional converage to parts of the api that weren't
present in older versions.