update gc logger to report using power of 1024 (#309)
The arguments to java for setting the heap sizes using
a power of 1024. This change updates the logger to report
using a power of 1024 instead of power of 1000 to match
the jvm behavior. To try and minimize confusion labels
use KiB, MiB, and GiB to make it explicit in the output.