Added spectator-web
This module adds a /spectator/metrics HTTP endpoint that returns a different
JSON encoding than spring boot's /metrics endpoint. The JSON encoding contains
a richer representation of the registry to enlighten external observers.

The endpoint supports some filtering mechanisms to restrict the domain of
metrics exposed if the caller is only interested in particular metrics.