Major overhaul and cleanup.
One of the sources of complexity before was the impression that there were
duplicate measurements. You had questioned that in your original review but
I had recalled seeing it. I ran some experiments and wasnt seeing it anymore
so I concluded that I must have seen it earlier do to a bug on my side.

Taking this out let me simplify things a bit. I also removed the
MeasurementFilter and added a predicate instead. The collectMeasurements
wasnt that interesting anymore so I removed it. I also removed my TagValue
class and use a subclass of BasicTag that adds the getters needed by Jackson.

I also removed the default tag filters since they probably arent that useful,
but I keep them as query parameters on the API because they are useful there.