add helper for fixed rate scheduling
Adds helper class adapted from internal util for scheduling
at a fixed rate, but skipping if the execution of a task
takes longers than the frequency. The scheduled executors
in the JDK will queue the results by default. For a number
of the reporting use-cases we want fixed rate behavior, but
do not want queueing. If it falls behind, then there is a
counter we can use to detect it.