initial version of Atlas registry
An initial version of a registry that talks to Atlas
directly rather doing it via Servo. This allows apps
that do not need all of the legacy compatibility working
to shed many dependencies and simplify the usage.

This is still considered experimental for now and may
have changes to configuration and other settings.