remove tags with empty values
Atlas does not allow empty values. For the common tags
where the value often comes from an environment variable
empty values are sometimes used instead of not having
the environment variable. Unfortunately that breaks the
fallback logic in typesafe config. So the agent will now
filter out the empty values.