cleanup dependencies
Cleans up some of the dependencies based on the unused
dependency rule for the gradle-lint-plugin:

https://github.com/nebula-plugins/gradle-lint-plugin/wiki/Unused-Dependency-Rule

The lint plugin was not left in the build because it
has spurious warnings about some slf4j test dependencies
that are needed.

Should replace #335 for tightening up the dependencies
needed for spring. The summary of the report is below
(reformatted a bit to improve readability of commit
message):

```
✖ spectator-web-spring/build.gradle: 5 problems (0 errors, 5 warnings)

one or more classes are required by your code directly for
- org.springframework.boot:spring-boot-autoconfigure:1.2.8.RELEASE
- org.springframework:spring-beans:4.1.9.RELEASE
- org.springframework:spring-web:4.1.9.RELEASE

these dependencies should be removed since their artifacts are empty
- org.springframework.boot:spring-boot-starter-actuator:1.2.8.RELEASE
- org.springframework.boot:spring-boot-starter-data-rest:1.2.8.RELEASE
```