add metric for rough estimate of clock skew (#341)
This isn't super accurate, but it is a simple way to
detect if there is a large difference between the time
sources for the client and the server. In particular,
in some environments we saw NTP fail and over several
days some instances would be off by 45s or more.