make gauges more consistent with other basic types (#353)
Changes the registry to support a `gauge(Id)` that returns
a settable gauge value. This allows the underlying type to
be activity based like the other core types. With this change
gauges are no longer a special case in terms of how they are
supported internally.

The old methods still works and will get polled via a
background thread. This has the advantage that gauges based
on user registered functions are now isolated and do not
impact regular collection.

This change should enable some additional cleanup that will
be done in subsequent PRs.