ensure gauge polling stops if registry is reaped (#354)
When the registry gets garbage collected the task for
polling the gauges will now get stopped. As part of
this change CompositeRegistry no longer uses an instance
of DefaultRegistry for gauges.

Updates to findbugs 3.0.1 to pick up a fix for private
methods only used via lambdas getting flagged as unused.