use volatile for previous value on StepValue impl
Updates StepValue implementations to use a volatile
variable rather than AtomicLong for the previous. We
only use get and set so this is an easy way to reduce
a bit of memory overhead. Not a big difference, but it
is a trivial change and there are often many instances
of these classes.