add support for Atlas query and data expressions
Adds support for basic Atlas expressions to be used with
local filtering and aggregation. For some later work with
streaming data off this will form a minimal basis that can
be done locally on each node to reduce the volume going
out to only what is needed.

For example, assume we have a metric `requestsPerSecond`
that has dimensions for `status` with cardinality around 4,
`endpoint` with cardinality 20, and `country` with a
cardinality of about 250. So the overall number of values
is around `4 * 20 * 250 = 20,000`. If we want to stream
a query like:

```
name,requestsPerSecond,:eq,:sum
```

This will allow a local aggregation on the node to reduce
it to a single value rather than sending all 20k and doing
all reduction server side. This is similar to map and combiner
steps of a Hadoop style job.