fix gauges for servo registry
The `ServoGauge` class was not overriding `Gauge.set`
which by default does nothing. So updates to the
gauge were getting lost. In a later PR may remove the
default impl in the base interface to avoid similar
problems with other registry implementations in the
future. The default is only there to avoid breaking
backwards compatibility due to adding a new method
to the interface.