improve totalTime estimate for cass latencies (#362)
For cassandra 2.x the metrics 2.2 library is used. When
extracting latencies for sending to Atlas, the `totalTime`
was being estimated by using the `Mean` value reported via
jmx. However, the `Mean` in metrics 2.2 does not come from
the snapshot used for the percentiles, but is since the timer
was created. This means that it isn't very reflective of
recent behavior. Metrics 3 changes that so that the `Mean`
is useful and is used with cassandra 3.x.

For now since we need to work with cassandra 2.x, the config
for `totalTime` uses the `50thPercentile` reported in jmx to
get the estimate.