zero metrics from snapshot if no change in count (#366)
When mapping in metrics from cassandra and possibly
other sources using metrics2/metrics3, the histograms
are based on a reservoir that is only rescaled during
updates. This can give a false impression of incorrect
sustained activity because the old value will keep
getting reported if no updates are received on the
underlying timer.

This change modifies the expressions to check that
the `Count` has changed since the previous time the
values were sampled. If the count has not changed,
then there was no activity so we zero out those values.
Otherwise the sampled value is returned without
modification.