A counter that keeps track of its last update
This adds ActivityCounter: a counter that keeps track of the last time
it was incremented. This fixes #369

Currently this is added to spectator-api as a first approach. It follows
the same convention of the Percentile meters where the factory method
takes a Registry as a parameter to avoid polluting the registry with
extra methods.