Rename to IntervalCounter and deal with expirations
IntervalCounter should make it clear that this Meter provides
more than just a Counter. It also provides the interval since the
last update. Since the main use case for this is when there's
a long period of inactivity that must be caught, it makes sense
to avoid the normal expiration of Meters, and keep reporting
the age.