add map for storing arbitrary state with registry
For long task timer (#237), interval counter (#369), and some
other uses it is necessary to maintain some state that is used
to provide values for the core types. The lifecycle of the
state should typically be tied to the lifecycle of the registry.

For example, the long task timer needs to keep track of tasks
and it is helpful if that will work across fetches of the timer
from the registry (#237).

This change introduces a concurrent map that can be accessed
from the registry to allow these composite helper types to
manage that state, but have the lifecycle of the objects tied
to the registry. That is, it will go away when the registry
goes away unless the user is maintaining references to it
elsewhere.

`LongTaskTimer` has been refactored to use this state as an
example and moved to the `patterns` package to keep this sort
of helper separate as we will likely add more over time. In
a future version the `Registry.longTaskTimer` will probably
get deprecated so that `LongTaskTimer` is not longer a special
case.