always increment count for recorded values
The Atlas registry was only incrementing the timer and
distribution summary objects if the recorded amount was
positive. If the amount was zero or negative it was
ignored.

Negative latency values are sometimes seen due to clock
adjustments on a machine. Also really quick events that
call `System.currentTimeMillis` for getting the time can
get a latency of 0. These should be reflected in the count
statistic for the timer or else it causes confusion
because the number of events will not match what the user
expects.

With this change negative and 0 values will update the
count, but not other statistics.