add config option to list dependency jars (#377)
Since the agent is normally used as a fat jar, it can
be hard to debug and figure out exactly what version is
being used. With this change a property in the typesafe
config, `netflix.spectator.agent.dependencies`, can be
used to specify a list of dependency jars that will be
exported to system properties by the agent. This allows
them to be easily queried using JMX.