add config options to set the allowed characters (#378)
Adds options to Atlas config interface for setting the
allowed characters of tag keys and to support overrides
for some tag values. The main use-case is allowing some
group names to include `^` and `~` as was recently agreed
internally.

The character set aspect has been moved to a utility
class `AsciiSet` with some performance tuning. In particular
the following changes were made:

1. The string is checked first and a buffer is only allocated
   if there are invalid characters. For the expected common
   case where there are no changes this gives a speed up of
   about 4-5x. In benchmarks it was about 15% slower for
   the case where changes are needed.
2. Switched from `StringBuilder` to using a `char[]` for the
   buffer. This is about 3x faster.
3. After 2, we have an array with the exact string so it can
   be used directly for the string to avoid an allocation and
   copy of the final string object. In the included benchmarks
   this was about 25% faster for the cases where changes were
   needed.