fix #312, preserve core types with Composite (#381)
The composite registry now preserves the core types
when using `get` as well as the stream methods. It
was broken before because `get` was just wrapping it
as a `CompositeMeter`. With this change `get` will now
inspect the type from the sub-registries and use the
type of the first one it finds.