move AtlasRegistry POJOs to impl package (#386)
There are some use-cases where it would be useful to
experiment with these and reuse the classes. However,
we want the flexibility to change them on a whim without
worrying about backwards compatibility. The compromise
for now is to move them to the `impl` package and make
them public with the clear expectation that any use
outside of spectator may break and that we are making
no commitment of backwards compatibility for those
classes.