explicitly specifiy DataSourceType.GAUGE for servo (#388)
Updates the gauges using ServoRegistry to explicity specify
`DataSourceType.GAUGE`. The Counter, Timer, and
DistributionSummary types get it automatically from using
StepCounter and MaxGauge internally.

This fixes the reporting of gauges when local
normalization is on. By default the value would get
treated as a rate and use a weighted average which
doesn't make sense for gauges.