fix #391, ClassCastException with :in
The same variable was being used for keeping track
of a list being created from the stack and popping
a list value off the stack to use for an operator.
The result is that operators like `:in` that take
a list would change the parse to list mode after
popping off the list argument and all subsequent
values on the stack would get added to the list until
another list terminator was found or the end of the
expression was reached.