fix #392, multiple ServoRegistry instances (#400)
Though not advised, this change will allow multiple
ServoRegistry instances to be created without clobbering
each other. If users are relying on this behavior their
code is broken and they should fix it to properly inject
the registry or use the global registry. However, with
this change it will prevent arbitrary broken behavior
from a library from breaking all spectator reporting
when it relies on the ServoRegistry.