reduce allocations for common withTags calls
Though there are typically better options, the variable
args `withTags(String... tags)` call is quite popular. This
change reduces the number of allocations for a number of
common calling patterns. In particular when using 4 or fewer
tags there are now overloads that will avoid the parameter
array from being created. Also if there are no existing
tags to merge with, it will dedup using the existing tags
array instead of allocating an new one can merging in.