AbstractRegistry now returns swappable impl
The AbstractRegistry which is the basis for most of the
other implementations now returns swappable implementations
of the core types. This is a first step that will allow
us to perform some additional steps later.

Why is this needed? Spectator allows the user to keep a
reference to a meter so it can be quickly updated without
an additional lookup cost. The drawback is that we cannot
safely expire the meter in the registry if someone is holding
onto a reference. If the user is running with improper tag
values this can grow over time and result in a memory leak.
For apps that pull in many libraries expecting each dependency
to be responsible with metrics is a tall order. This change
will allow us to swap out the reference when expiring to
a simple implementation that will lookup again if there is
activity.

In addition, this should give us the ability to due more
adaptive rollup work (#22) in the future.