add support for aws java sdk 2 (#451)
Adds an extension that works with the latest
developer preview of the 2.x version of the sdk.
This is experimental until there is a final 2.x
release, but allows for some experimentation with
the developer preview at Netflix.

The aws sdk depends on jackson-jr which is effected
by a bug in jdk9 (FasterXML/jackson-jr#55). The
travis build has been updated to allow the jdk9 build
to fail. When there is a jackson-jr 2.9.1 available
with a fix, then that can be changed.