remove default impl for Registry.gauge (#455)
The default implementation that does nothing was added
to avoid breaking compatibility for external classes
that implementated that interface. However, that comes
at the cost of correctness if that method is not
implemented. Current estimate is that it is low risk
to remove that default and it makes the api cleaner
and less error prone.

It would also be nice to remove the default implementation
for `Gauge.set`, but there appear to be some uses and we
want to avoid heavily breaking changes at this time.