rename passive gauge overloads
Deprecates the passive overloads that used to be named
`gauge` and replaces them with `monitorNumber` and
`monitorValue` methods. The difference between passive
and active gauge has caused some confusion and the goal
is to make the distinction more clear while maintaining
consistent naming for the core active types.

The methods have been deprecated with a note that they
will be removed in 2.0. For now the desire is to avoid
significantly breaking backwards compatibility with
the api between now and marking as 1.0.