move polled gauges to patterns package
The polled gauges are just a pattern for something that is
regularly polled in the background and updates an active
gauge type on the registry. One common source of confusion
is the registry class has gauge overloads for both the
active and passive variants. With this change the passive
variants have been deprecated on the Registry interface
and the PolledGauge class can be used to support the
passive use-cases.