fix NPE for error responses with empty body
If an error response such as a 503 is returned, then
HttpURLConnection will return `null` for the input
stream of the body. This change now checks for that
and returns an empty array rather than failing with:

```
java.lang.NullPointerException
	at com.netflix.spectator.sandbox.HttpRequestBuilder.readAll(HttpRequestBuilder.java:264)
	at com.netflix.spectator.sandbox.HttpRequestBuilder.sendImpl(HttpRequestBuilder.java:242)
	at com.netflix.spectator.sandbox.HttpRequestBuilder.send(HttpRequestBuilder.java:165)
	at com.netflix.spectator.sandbox.DefaultHttpClientTest.unavailableEmpty(DefaultHttpClientTest.java:127)
```

The `DefaultHttpClientTest.unavailableEmpty()` test
case that was added verifies this particular change.