use eager singleton for Registry binding (#474)
This is mostly to help avoid issues with some internal
users that are using spring along with a bridge that
brings in guice bindings. If the bindings are lazy, then
apparently the bridge cannot tell the type for the binding
and use the concrete type of the class. For the registry
this means it cannot tell there is already a binding for
the Registry interface because it just sees the concrete
ServoRegistry type.