fix #483, ArrayTagSet.dedup overwrites first value (#484)
There was a bug in ArrayTagSet.dedup where if the source
list had multiple entries that are not duplicates, then
it would overwrite the value of the first entry with the
key of the second. This would lead to a shifted set of
tags and the last value is null. This causes a NPE when
attempting to compute the hashCode.