gradle 4.4 and nebula 5 (#506)
No major improvements we care about, but keeps us in
sync with versions recommended by engtools.

For some reason, this version of gradle on jdk9 makes it
much more likely that a GC will happen between a gauge
being registered and the value getting sampled. This
exposed a bug in the compatibility test so that jar has been
updated with the patch from brharrington/spectator-examples#4.