Add gauge to Scheduler to record the timestamp of the last execution.
This commit adds monitoring of when the last execution occurred and tags
it with the result (failure if an exception was thrown, success
otherwise).  We had a case where a scheduled thread stopped running at
some point, but had no indication due to the nature of the work it does.
This will enables a "time since last success" alert.