make Scheduler robust to uncaught Throwables (#509)
If a Throwable was propagated from an executed task,
then the thread would get stopped and never replaced.
For many uses of the scheduler, we cannot control
the task to ensure they are well-behaved and do not
want a bad task to impact other tasks that may be on
the same scheduler. With this change we catch Throwable.

An additional counter is added to track the number
of uncaught exceptions that occur for the scheduler.
This will make it easy to alert if we are seeing
regular failures.

To simplify updating the stats without having to pass
around too many handles, they have been consolidated
in a stats class.

There is also a small fix to avoid having threads
restart after shutdown. Before the shutdown was
interrupting the threads and a new one would be
created to replace them. So there would still be
threads running for the scheduler after after it
was shutdown.