fix #501, change package for spectator-nflx
The package name for the module annotated to be an
auto-bind singleton has been moved so it does not
conflict with `spectator-nflx-plugin`. This module
was changed for two reasons:

1. It is deprecated and if we have failures for users
   we would prefer it be on this library so we can
   move users off.
2. This class is only loaded via classpath scanning
   so it shouldn't break anything.

Note, the test case has been left in the old package
because it needs package private access to the Plugin
class. Since this is internal to the code base we'll
revisit if it becomes a problem later. External users
will not be impacted.