add flag to ignore the query when aggregating
Previously, when aggegating values for a DataExpr the
query would get checked for every update. For some
use-cases the query will already have been validated
and this causes them to get checked twice adding a
significant overhead.

In addition, the group by and rollup operators now
use that for the grouped results to avoid rechecking
the query internally.