add support for removing expired meters (#537)
Originally the contract for the registry was that
the meters like counter would track the state since
the meter was created. This means that meters must
stick around for the life of the process.

With some registries, e.g., AtlasRegistry, we already
removed this retriction and limited it to the count
for the last completed interval. Given we have been
running in this mode a while with no issues, we are
changing the contract to officially give the registry
implementation the choice of reset frequency. Along
with this the AbstractRegistry now has a method that
can be used to fully remove any expired meters. This
avoids a memory leak over time if the user has a bug
and the distinct meter set grows over time.