ignore "name" when used as a tag key (#538)
For Atlas the name will be filled in by the name
on the id. If the user used this key, then ignore
it when sending to Atlas. Otherwise it could overwrite
the name on the id depending on how the json map
is consumed.

In a separate PR we'll add some validation logging
so the user can be aware if they are doing something
wrong.