fix possible missed timer updates using lambda
The SwapTimer implementation allows us to swap or null out
the underlying timer implementation and recreate it if the
user still had a reference somewhere and there is new
activity.

There was a bug in the record calls that took a lambda where
it would cache the underlying timer before the lambda was
called and reuse it again after. For a lambda with a longer
duration, it was quite likely the underlying timer would be
removed before this completed and the update would get dropped.

This change also modifies the initial state of the meters when
first created to not be expired. Before they were initially
expired until the first activity occurred.