add max gauge type to registry (#548)
This was added a while back to the AtlasRegistry for
use in the aggregator service and is used internally
to timer and distribution summary implementations. To
make things more consistent it is being added to the
main registry interface.

Travis build switched to use container rather than
dedicated GCE instance because of network problems
on Travis. This can be reverted after the issue is
fixed if it is causing problems. So far the build seems
to work fine on the container.