fix secondary expiration of cached meter (#549)
If there was a cached reference to a meter and it expired
more than once, then it would be updating a dangling meter
that is no longer connected to the registry.

The first expiration would work because the SwapMeter in
the registry and the SwapMeter referenced by the user are
the same. After that, the SwapMeter in the registry is
different and so it would not null out the value for the
cached copy and force a new lookup. This has been simplified
to just check for expiration of the underlying meter and
perform a lookup if expired.